refactor(mcp): share trends line chart via a dumb view component
Extract a dependency-clean presentational `TrendsLineChartView` so the MCP trends
line chart and the web chart can share one render path instead of duplicating
the series + derived-config assembly.

- `trendsSeriesTransforms.ts` (new): the framework-agnostic transforms
  (`buildTrendsSeries`, `buildDerivedConfigs`, `computeDashedFromIndex`,
  `TrendsResultLike`) split out of `trendsChartTransforms.ts` with only an
  `@posthog/quill-charts` dependency, so the MCP UI app build can bundle it.
  `trendsChartTransforms.ts` re-exports them, so existing callers are unchanged.
- `TrendsLineChartView.tsx` (new): dumb component — takes results + display
  options, builds series/derived configs, renders `TimeSeriesLineChart`. Accepts
  framework-specific bits (axis, value labels, goal lines, tooltip, click,
  children) as props so the web chart can adopt it later.
- `McpTrendsLineChart` collapses to a thin adapter that maps MCP results to the
  view's props (palette, label fallback, y-unit, interval/timezone).

Also hardens the shared transforms for `exactOptionalPropertyTypes` /
`noUncheckedIndexedAccess` (enabled in the MCP build but not the web build), which
also fixes the palette index access in the MCP adapter.

Generated-By: PostHog Code
Task-Id: 6f4c75d8-df09-4a9d-a11f-22435268e155
